Extract flow state into useFlowStore hook
Created src/hooks/useFlowStore.js owning all flow data:
- 17 content state items (transcript, parsed, xml, improvements, metrics, etc.)
- Flows list + currentFlowId with localStorage persistence
- Auto-save, migration, and restore-on-mount effects
- Flow navigation: handleOpenFlow, handleCreateFlow, handleBackToDashboard,
  handleClearCurrentFlow, handleDeleteFlow, handleExported

App.jsx now destructures everything from useFlowStore({ vimplToken, vimplUser }).
Retains only UI state (loading flags, modal visibility, activePanel).

Closes #1
